requirements.txt
Django==4.0
djangorestframework==3.13.1
requests==2.26.0
django-storages==1.12.3
boto3~=1.20.41
drf-nested-routers==0.93.4
gunicorn==20.1.0
psycopg2==2.9.3
dj-config-url~=0.1.1
django-encrypted-model-fields==0.6.1
ics~=0.7